[REF] *: cleanup legacy test helpers
*calendar,hr_attendance,hr_calendar,l10n_ro_edi,mail,mrp,web,
website,website_livechat

Now that almost all qunit tests have been converted to hoot, we
can get rid of a lot of legacy helpers. This PR only aims at
removing what is no longer used, or what can be trivially
substituted. A further work will be necessary as some legacy
helpers are used in tours. Moreover, the whole QUnit test suite
system will still have to be removed, once the few remaining tests
will be converted.
