+++ addons/hr_attendance/__manifest__.py
@@ -39,9 +39,6 @@ actions(Check in/Check out) performed by them.
             'hr_attendance/static/src/**/*.xml',
             'hr_attendance/static/src/scss/views/*.scss'
         ],
-        'web.qunit_suite_tests': [
-            'hr_attendance/static/tests/hr_attendance_mock_server.js',
-        ],
         'hr_attendance.assets_public_attendance': [
             # Define attendance variables (takes priority)
             'hr_attendance/static/src/scss/kiosk/primary_variables.scss',
