+++ addons/mrp/static/tests/mrp_document_kanban.test.js
@@ -6,9 +6,9 @@ import {
     start,
     startServer
 } from "@mail/../tests/mail_test_helpers";
+import { inputFiles } from "@mail/../tests/mail_test_helpers_contains";
 import { defineMrpModels } from "@mrp/../tests/mrp_test_helpers";
 import { describe, test } from "@odoo/hoot";
-import { inputFiles } from "@web/../tests/utils";
 import { asyncStep, getService, patchWithCleanup, waitForSteps } from "@web/../tests/web_test_helpers";
 import { fileUploadService } from "@web/core/file_upload/file_upload_service";
