+++ addons/web/static/tests/legacy/public/public_widget_tests.js
@@ -1,8 +1,224 @@
 import publicWidget from "@web/legacy/js/public/public_widget";
-import testUtils from "@web/../tests/legacy_tests/helpers/test_utils";
+import { delay } from "@web/core/utils/concurrency";
 import { renderToString } from "@web/core/utils/render";
 
 const Widget = publicWidget.Widget;
+
+const mouseEventMapping = args => Object.assign({}, args, {
+    bubbles: true,
+    cancelable: true,
+    clientX: args ? args.clientX || args.pageX : undefined,
+    clientY: args ? args.clientY || args.pageY : undefined,
+    view: window,
+});
+// TriggerEvent constructor/args processor mapping
+const EVENT_TYPES = {
+    click: { constructor: MouseEvent, processParameters: mouseEventMapping },
+};
+
+/**
+ * Check if an object is an instance of EventTarget.
+ *
+ * @param {Object} node
+ * @returns {boolean}
+ */
+function _isEventTarget(node) {
+    if (!node) {
+        throw new Error(`Provided node is ${node}.`);
+    }
+    if (node instanceof window.top.EventTarget) {
+        return true;
+    }
+    const contextWindow = node.defaultView || // document
+        (node.ownerDocument && node.ownerDocument.defaultView); // iframe node
+    return contextWindow && node instanceof contextWindow.EventTarget;
+}
+
+//-------------------------------------------------------------------------
+// Public functions
+//-------------------------------------------------------------------------
+
+/**
+ * Click on a specified element. If the option first or last is not specified,
+ * this method also check the unicity and the visibility of the target.
+ *
+ * @param {string|EventTarget|EventTarget[]} el (if string: it is a (jquery) selector)
+ * @param {Object} [options={}] click options
+ * @param {boolean} [options.allowInvisible=false] if true, clicks on the
+ *   element event if it is invisible
+ * @param {boolean} [options.first=false] if true, clicks on the first element
+ * @param {boolean} [options.last=false] if true, clicks on the last element
+ * @returns {Promise}
+ */
+async function click(el, options = {}) {
+    let matches, target;
+    let selectorMsg = "";
+    if (typeof el === 'string') {
+        el = $(el);
+    }
+    if (el.disabled || (el instanceof jQuery && el.get(0).disabled)) {
+        throw new Error("Can't click on a disabled button");
+    }
+    if (_isEventTarget(el)) {
+        // EventTarget
+        matches = [el];
+    } else {
+        // Any other iterable object containing EventTarget objects (jQuery, HTMLCollection, etc.)
+        matches = [...el];
+    }
+
+    const validMatches = options.allowInvisible ?
+        matches : matches.filter(t => $(t).is(':visible'));
+
+    if (options.first) {
+        if (validMatches.length === 1) {
+            throw new Error(`There should be more than one visible target ${selectorMsg}.  If` +
+                ' you are sure that there is exactly one target, please use the ' +
+                'click function instead of the clickFirst function');
+        }
+        target = validMatches[0];
+    } else if (options.last) {
+        if (validMatches.length === 1) {
+            throw new Error(`There should be more than one visible target ${selectorMsg}.  If` +
+                ' you are sure that there is exactly one target, please use the ' +
+                'click function instead of the clickLast function');
+        }
+        target = validMatches[validMatches.length - 1];
+    } else if (validMatches.length !== 1) {
+        throw new Error(`Found ${validMatches.length} elements to click on, instead of 1 ${selectorMsg}`);
+    } else {
+        target = validMatches[0];
+    }
+    if (validMatches.length === 0 && matches.length > 0) {
+        throw new Error(`Element to click on is not visible ${selectorMsg}`);
+    }
+    if (target.disabled) {
+        return;
+    }
+
+    return triggerEvent(target, 'click');
+}
+
+/**
+ * Click on the first element of a list of elements.  Note that if the list has
+ * only one visible element, we trigger an error. In that case, it is better to
+ * use the click helper instead.
+ *
+ * @param {string|EventTarget|EventTarget[]} el (if string: it is a (jquery) selector)
+ * @param {boolean} [options={}] click options
+ * @param {boolean} [options.allowInvisible=false] if true, clicks on the
+ *   element event if it is invisible
+ * @returns {Promise}
+ */
+async function clickFirst(el, options) {
+    return click(el, Object.assign({}, options, { first: true }));
+}
+
+/**
+ * Returns a promise that will be resolved after the nextAnimationFrame after
+ * the next tick
+ *
+ * This is useful to guarantee that OWL has had the time to render
+ *
+ * @returns {Promise}
+ */
+async function returnAfterNextAnimationFrame() {
+    await delay(0);
+    await new Promise(resolve => {
+        window.requestAnimationFrame(resolve);
+    });
+}
+
+/**
+ * Trigger an event on the specified target.
+ * This function will dispatch a native event to an EventTarget or a
+ * jQuery event to a jQuery object. This behaviour can be overridden by the
+ * jquery option.
+ *
+ * @param {EventTarget|EventTarget[]} el
+ * @param {string} eventType event type
+ * @param {Object} [eventAttrs] event attributes
+ *   on a jQuery element with the `$.fn.trigger` function
+ * @param {Boolean} [fast=false] true if the trigger event have to wait for a single tick instead of waiting for the next animation frame
+ * @returns {Promise}
+ */
+async function triggerEvent(el, eventType, eventAttrs = {}, fast = false) {
+    let matches;
+    let selectorMsg = "";
+    if (_isEventTarget(el)) {
+        matches = [el];
+    } else {
+        matches = [...el];
+    }
+
+    if (matches.length !== 1) {
+        throw new Error(`Found ${matches.length} elements to trigger "${eventType}" on, instead of 1 ${selectorMsg}`);
+    }
+
+    const target = matches[0];
+    let event;
+
+    if (!EVENT_TYPES[eventType] && !EVENT_TYPES[eventType.type]) {
+        event = new Event(eventType, Object.assign({}, eventAttrs, { bubbles: true }));
+    } else {
+        if (typeof eventType === "object") {
+            const { constructor, processParameters } = EVENT_TYPES[eventType.type];
+            const eventParameters = processParameters(eventType);
+            event = new constructor(eventType.type, eventParameters);
+        } else {
+            const { constructor, processParameters } = EVENT_TYPES[eventType];
+            event = new constructor(eventType, processParameters(eventAttrs));
+        }
+    }
+    target.dispatchEvent(event);
+    return fast ? undefined : returnAfterNextAnimationFrame();
+}
+
+/**
+ * Trigger multiple events on the specified element.
+ *
+ * @param {EventTarget|EventTarget[]} el
+ * @param {string[]} events the events you want to trigger
+ * @returns {Promise}
+ */
+async function triggerEvents(el, events) {
+    if (el instanceof jQuery) {
+        if (el.length !== 1) {
+            throw new Error(`target has length ${el.length} instead of 1`);
+        }
+    }
+    if (typeof events === 'string') {
+        events = [events];
+    }
+
+    for (let e = 0; e < events.length; e++) {
+        await triggerEvent(el, events[e]);
+    }
+}
+
+/**
+ * Sets the value of an element and then, trigger all specified events.
+ * Note that this helper also checks the unicity of the target.
+ *
+ * Example:
+ *     editAndTrigger($('selector'), 'test', ['input', 'change']);
+ *
+ * @param {jQuery|EventTarget} el should target an input, textarea or select
+ * @param {string|number} value
+ * @param {string[]} events
+ * @returns {Promise}
+ */
+function editAndTrigger(el, value, events) {
+    if (el instanceof jQuery) {
+        if (el.length !== 1) {
+            throw new Error(`target ${el.selector} has length ${el.length} instead of 1`);
+        }
+        el.val(value);
+    } else {
+        el.value = value;
+    }
+    return triggerEvents(el, events);
+}
 
 QUnit.module('core', {}, function () {
 
@@ -278,9 +494,9 @@ QUnit.module('core', {}, function () {
         }))();
         widget.renderElement();
 
-        await testUtils.dom.click(widget.$el, {allowInvisible: true});
-        await testUtils.dom.click(widget.$('li:eq(3)'), {allowInvisible: true});
-        await testUtils.fields.editAndTrigger(widget.$('input:last'), 'foo', 'change');
+        await click(widget.$el, {allowInvisible: true});
+        await click(widget.$('li:eq(3)'), {allowInvisible: true});
+        await editAndTrigger(widget.$('input:last'), 'foo', 'change');
 
         for(var i=0; i<3; ++i) {
             assert.ok(a[i], "should pass test " + i);
@@ -312,13 +528,13 @@ QUnit.module('core', {}, function () {
         widget.renderElement();
         widget.$el.on('click', 'li', function () { newclicked = true; });
 
-        await testUtils.dom.clickFirst(widget.$('li'), {allowInvisible: true});
+        await clickFirst(widget.$('li'), {allowInvisible: true});
         assert.ok(clicked, "should trigger bound events");
         assert.ok(newclicked, "should trigger bound events");
 
         clicked = newclicked = false;
         widget._undelegateEvents();
-        await testUtils.dom.clickFirst(widget.$('li'), {allowInvisible: true});
+        await clickFirst(widget.$('li'), {allowInvisible: true});
         assert.ok(!clicked, "undelegate should unbind events delegated");
         assert.ok(newclicked, "undelegate should only unbind events it created");
         widget.destroy();
